# src/main/java/it/tim/bl/visualizza/richiesta/domfisso/bean/RequestBVRD.java
package it.tim.bl.visualizza.richiesta.domfisso.bean;

public class RequestBVRD {

	private VisualizzazioneRichiestaAttivazioneDomiciliazioneFisso visualizzazioneRichiestaAttivazioneDomiciliazioneFisso;

	public VisualizzazioneRichiestaAttivazioneDomiciliazioneFisso getVisualizzazioneRichiestaAttivazioneDomiciliazioneFisso() {
		return visualizzazioneRichiestaAttivazioneDomiciliazioneFisso;
	}

	public void setVisualizzazioneRichiestaAttivazioneDomiciliazioneFisso(
			VisualizzazioneRichiestaAttivazioneDomiciliazioneFisso visualizzazioneRichiestaAttivazioneDomiciliazioneFisso) {
		this.visualizzazioneRichiestaAttivazioneDomiciliazioneFisso = visualizzazioneRichiestaAttivazioneDomiciliazioneFisso;
	}


	
	
}